refactor: hoist duplicated interop helpers into shared modules (#1471)
async_sleep_ms, focus_*_panel, and now_unix were each independently
reimplemented in 5-6 frontend files. Hoist each into one shared module
(platform_sleep, focus_after_paint, time) and repoint every call site
at it — a mechanical dedupe with no behavior change.
