+++ frontend/src/pages/book_detail/rating.rs
@@ -8,6 +8,7 @@
 use dioxus::prelude::*;
 use omnibus_shared::{AttributedRating, RatingRecord, RatingUpdate};
 
+use crate::time::now_unix;
 use crate::{data, use_server_url};
 
 /// Live rating signals shared between [`BdRatingWidget`] and each
@@ -347,23 +348,6 @@ fn rated_ago(now: i64, updated_at: i64) -> String {
     } else {
         let d = secs / 86_400;
         format!("rated {d} day{} ago", plural(d))
-    }
-}
-
-/// Current unix time in seconds. Only ever called client-side (the relative
-/// timestamp renders after the post-mount load, and the optimistic write runs
-/// on click), so SSR never invokes the JS clock.
-fn now_unix() -> i64 {
-    #[cfg(feature = "web")]
-    {
-        (js_sys::Date::now() / 1000.0) as i64
-    }
-    #[cfg(not(feature = "web"))]
-    {
-        std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_secs() as i64)
-            .unwrap_or(0)
     }
 }
